fix: moved kafka code to prevent usage in non-kafka projects
