+++ golang/cmd/kafka-debug/kafka.go
@@ -2,16 +2,16 @@ package main
 
 import (
 	"github.com/confluentinc/confluent-kafka-go/kafka"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/internal"
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/kafka_helper"
 	"go.uber.org/zap"
 	"time"
 )
 
 func startDebugger() {
-	internal.KafkaConsumer.Subscribe("^ia.+", nil)
+	kafka_helper.KafkaConsumer.Subscribe("^ia.+", nil)
 	for !ShuttingDown {
 
-		msg, err := internal.KafkaConsumer.ReadMessage(5) //No infinitive timeout to be able to cleanly shut down
+		msg, err := kafka_helper.KafkaConsumer.ReadMessage(5) //No infinitive timeout to be able to cleanly shut down
 		if err != nil {
 			if err.(kafka.Error).Code() == kafka.ErrTimedOut {
 				continue
