+++ golang/cmd/kafka-init/kafka.go
@@ -1,14 +1,14 @@
 package main
 
 import (
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/internal"
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/kafka_helper"
 	"strings"
 )
 
 func initKafkaTopics(topics string) {
 	topiclist := strings.Split(topics, ";")
 	for _, topic := range topiclist {
-		err := internal.CreateTopicIfNotExists(topic)
+		err := kafka_helper.CreateTopicIfNotExists(topic)
 		panic(err)
 	}
 }
