+++ golang/cmd/kafka-init/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/confluentinc/confluent-kafka-go/kafka"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/internal"
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/kafka_helper"
 	"go.uber.org/zap"
 	"os"
 	"os/signal"
@@ -24,7 +24,7 @@ func main() {
 	KafkaTopics := os.Getenv("KAFKA_TOPICS")
 
 	zap.S().Debugf("Setting up Kafka")
-	internal.SetupKafka(kafka.ConfigMap{
+	kafka_helper.SetupKafka(kafka.ConfigMap{
 		"bootstrap.servers": KafkaBoostrapServer,
 		"security.protocol": "plaintext",
 		"group.id":          "kafka-init",
@@ -60,7 +60,7 @@ func main() {
 func ShutdownApplicationGraceful() {
 	zap.S().Infof("Shutting down application")
 
-	internal.CloseKafka()
+	kafka_helper.CloseKafka()
 
 	zap.S().Infof("Successfull shutdown. Exiting.")
