+++ golang/cmd/mqtt-kafka-bridge/main.go
@@ -6,6 +6,7 @@ import (
 	MQTT "github.com/eclipse/paho.mqtt.golang"
 	"github.com/heptiolabs/healthcheck"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/internal"
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/kafka_helper"
 	"go.uber.org/zap"
 	"net/http"
 	"os"
@@ -69,12 +70,12 @@ func main() {
 	SetupMQTT(MQTTCertificateName, MQTTBrokerURL, MQTTTopic, health, podName, mqttIncomingQueue)
 
 	zap.S().Debugf("Setting up Kafka")
-	internal.SetupKafka(kafka.ConfigMap{
+	kafka_helper.SetupKafka(kafka.ConfigMap{
 		"bootstrap.servers": KafkaBoostrapServer,
 		"security.protocol": "plaintext",
 		"group.id":          "mqtt-kafka-bridge",
 	})
-	err = internal.CreateTopicIfNotExists(KafkaBaseTopic)
+	err = kafka_helper.CreateTopicIfNotExists(KafkaBaseTopic)
 	if err != nil {
 		panic(err)
 	}
@@ -116,7 +117,7 @@ func ShutdownApplicationGraceful() {
 	ShuttingDown = true
 	mqttClient.Disconnect(1000)
 
-	internal.CloseKafka()
+	kafka_helper.CloseKafka()
 
 	time.Sleep(15 * time.Second) // Wait that all data is processed
