+++ golang/kafka_helper/kafka.go
@@ -1,4 +1,4 @@
-package internal
+package kafka_helper
 
 import (
 	"context"
